refactor(subtitle): move module Subtitle.js to Subtitle/index.js
